requirements-vercel.txt
fastapi==0.68.1
python-multipart==0.0.5
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
numpy==1.24.3
uvicorn==0.15.0 